requirements.txt
telebot
pyTelegramBotApi==3.6.6
bs4
lxml
requests
asyncio
aiohttp
pandas
openpyxl
mysql-connector-python